WIP: ikev2: Ignore roam events without MOBIKE but static local address
Disabling MOBIKE and statically configuring a local address should be
enough indication that the user doesn't want to roam to a different
address.  There might not be any routes that indicate we can use the
current address but it might still work (e.g. if the address is on an
interface that is not referenced in any routes and the address itself
is neither).  This way we avoid switching to another address for routes
that might be available on the system.

We currently don't make much use of COND_STALE anyway when MOBIKE is not
enabled, e.g. to avoid sending DPDs if the connection is seemingly down.
With MOBIKE enabled we don't exactly check that state but we do don't
send DPDs if there is no route/source address available.

WIP: Using match_me() that way is kind of a hack. Maybe adding a helper
function that only checks IP addresses (not hostnames, ranges and
subnets too) and does not return an arbitrary integer would make this
clearer and less prone to failure due to future refactorings that may
change the internals of that function.

Also, we could check if the current IP is still present on the system
and perhaps switch COND_STALE on if not (even though, as noted above, we
currently make not much use of this - the only thing is sending NAT-T
keepalives).